# STC and PSAR performance improvements