--- a/pandas_ta/momentum/stc.py
+++ b/pandas_ta/momentum/stc.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+import pandas as pd
+import numpy as np
 from numpy import nan
 from pandas import DataFrame, Series
 from pandas_ta._typing import DictLike, Int, IntFloat
@@ -153,18 +155,18 @@ def stc(
 
     return df
 
-
 def schaff_tc(close, xmacd, tclength, factor):
     # ACTUAL Calculation part, which is shared between operation modes
-    # 1St : Stochastic of MACD
-    lowest_xmacd = xmacd.rolling(tclength).min()  # min value in interval tclen
+    lowest_xmacd = xmacd.rolling(tclength).min()
     xmacd_range = non_zero_range(xmacd.rolling(tclength).max(), lowest_xmacd)
     m = len(xmacd)
 
-    # %Fast K of MACD
-    stoch1, pf = list(xmacd), list(xmacd)
-    stoch1[0], pf[0] = 0, 0
+    # Initialize lists
+    stoch1, pf = [0] * m, [0] * m
+    stoch2, pff = [0] * m, [0] * m
+
     for i in range(1, m):
+        # %Fast K of MACD
         if lowest_xmacd[i] > 0:
             stoch1[i] = 100 * ((xmacd[i] - lowest_xmacd[i]) / xmacd_range[i])
         else:
@@ -172,21 +174,26 @@ def schaff_tc(close, xmacd, tclength, factor):
         # Smoothed Calculation for % Fast D of MACD
         pf[i] = round(pf[i - 1] + (factor * (stoch1[i] - pf[i - 1])), 8)
 
-    pf = Series(pf, index=close.index)
+        # find min and max so far
+        if i < tclength:
+            # If there are not enough elements for a full tclength window, use what is available
+            lowest_pf = min(pf[:i+1])
+            highest_pf = max(pf[:i+1])
+        else:
+            lowest_pf = min(pf[i-tclength+1:i+1])
+            highest_pf = max(pf[i-tclength+1:i+1])
 
-    # 2nd : Stochastic of smoothed Percent Fast D, 'PF', above
-    lowest_pf = pf.rolling(tclength).min()
-    pf_range = non_zero_range(pf.rolling(tclength).max(), lowest_pf)
+        # Ensure non-zero range
+        pf_range = highest_pf - lowest_pf if highest_pf - lowest_pf > 0 else 1
 
-    # % of Fast K of PF
-    stoch2, pff = list(xmacd), list(xmacd)
-    stoch2[0], pff[0] = 0, 0
-    for i in range(1, m):
-        if pf_range[i] > 0:
-            stoch2[i] = 100 * ((pf[i] - lowest_pf[i]) / pf_range[i])
+        # % of Fast K of PF
+        if pf_range > 0:
+            stoch2[i] = 100 * ((pf[i] - lowest_pf) / pf_range)
         else:
             stoch2[i] = stoch2[i - 1]
-        # Smoothed Calculation for % Fast D of PF
         pff[i] = round(pff[i - 1] + (factor * (stoch2[i] - pff[i - 1])), 8)
 
-    return pff, pf
+    pf_series = Series(pf, index=close.index)
+    pff_series = Series(pff, index=close.index)
+
+    return pff_series, pf_series